requirements.txt
asgiref==3.8.1
Django==5.0.6
django-location-field==2.7.3
psycopg2-binary==2.9.9
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1